refactor(idea_service): extract write CRUD ops (#163)
  before: 1395
  after:  1007 (-388)
